svnrev: Produce a git version if we're in a git repo
We check for a readable `.git` to see if we're in a Git repo or not.
(Remember that this may be a file, as well as a directory.)

If we are, we use `git describe` to print the current short commit
hash with `-dirty` appended if our working copy isn't clean. This is
probably far from perfect, but good enough for the moment.