WIP Start in on more vi-ish commands
XXX Explain here what we've done.
XXX Update documentation.

We currently disable the old `X` and `O` commands, that do
"big/little-endian binary" as  e.g. `X_X_XX__` and `__OO_O_O` because
we need `O` for "open line above". We probably want to restore these
as part of a two char "binary" command that starts with `%` and then
takes a second char to determine which of the several binary formats
to use.