Rebind labelling, tracing, goto-ref/return, hint flags
We introduce the convention that lower-case commands `x` operate on
the current line ("_sel") and upper-case commands `X`operate on the
"refaddr," or the line which the current instruction's argument
references. This helps by freeing up ^L and ^T for their vi functions.

This introduces new single-key commands for setting labels for
_sel/refaddr; especially for the latter this is much faster and easier
than jumping to the refaddr, `:l ` to set the label, and then going
back.

The new bindings are:

     l   L   set the label for _sel/refaddr (new commands)
     t   T   trace disassembly for _sel/refaddr (was T,^T)
    ^]  ^T   go to refaddr; return from previous jump (was/is @,<)
         F   toggle hint flags (was ")
     '   "   change label type at _sel/refaddr (was ^,^L)
        ^L   refresh screen (new command)

Vi has no default key binding for jumping to a newer entry on the tag
stack, so we leave this as `>` here.

This code is admittedly a bit hacky and could use some cleanup, but I
think we need some general rework to, e.g., disconnect the command
function names from the names of the keys to which they are bound.

XXX None of this is documented yet.