update: Script to import new commits from the trunk of original SVN repo
This uses `git svn` instead of `svn2git` (which is what we originally
used) because `svn2git` assumes various annoying things about branch
names, such as having to use `master` instead of `main` etc.

A sort-of-consequence of this is that we update only `trunk`, and do not
add new release branches/tags.